demo/requirements.txt
streamlit==1.31.0
torch==2.7.0
torchvision==0.22.0
pillow==10.0.0